Added Disabled Future Scenes To Build Settings
- Added additional scenes to the Build Settings, but not active since they are not completed
- Learn to create a Development Debug Build with the following steps based on this tutorial (https://docs.unity3d.com/2022.3/Documentation/Manual/ManagedCodeDebugging.html),
(1) Go to Build Settings and toggle on the Development Build, then Script Debugging, then Wait For Managed Debugger if you want to run the build in Visual Studio.
(2) In Visual Studio, go to Debug > Attach Unity Debugger, then select the Editor for the current project.
(3) Set appropriate breakpoints such as Debug messages and then run the game. When the breakpoints are reached, go back to Visual Studio to see any errors and warning messages and then press the green Continue arrow.